backend/requirements.txt
fastapi
uvicorn
numpy
pandas
pandas-gbq
redis[hiredis]
tqdm
python-dotenv